requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
django-phonenumber-field==8.0.0
phonenumbers==8.13.45
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
